Split network module into its own `sfml-network` crate
