+++ sfml-network/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "sfml-network"
+version = "0.1.0"
+authors = ["Mika Attila <radiantstatue@gmail.com>"]
+
+[dependencies.sfml]
+path =".."
+
+[dependencies.csfml-network-sys]
+path = "../csfml-network-sys"
+
+[dependencies.csfml-system-sys]
+path = "../csfml-system-sys"
