+++ sfml-network/src/socket_status.rs
