+++ sfml-network/src/tcp_listener.rs
@@ -25,12 +25,12 @@
 
 use std::mem;
 
-use system::raw_conv::Raw;
-use network::{TcpSocket, SocketStatus, IpAddress};
+use sfml::system::raw_conv::Raw;
+use {TcpSocket, SocketStatus, IpAddress};
 
 use csfml_system_sys::sfBool;
 use csfml_network_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use sfml::system::SfBoolExt;
 
 /// Socket that listens to new TCP connections
 pub struct TcpListener {
