+++ sfml-network/src/udp_socket.rs
@@ -25,12 +25,12 @@
 
 use std::{ptr, mem};
 
-use system::raw_conv::{Raw, FromRaw};
-use network::{Packet, IpAddress, SocketStatus};
+use sfml::system::raw_conv::{Raw, FromRaw};
+use {Packet, IpAddress, SocketStatus};
 
 use csfml_system_sys::sfBool;
 use csfml_network_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use sfml::system::SfBoolExt;
 
 /// Specialized socket using the UDP protocol.
 pub struct UdpSocket {
