+++ src/audio/sound.rs
@@ -30,7 +30,7 @@ use system::raw_conv::{Raw, FromRaw};
 
 use csfml_system_sys::{sfBool, sfVector3f};
 use csfml_audio_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Regular sound that can be played in the audio environment.
 ///
