+++ src/audio/sound_buffer.rs
@@ -27,7 +27,7 @@ use system::raw_conv::{Raw, FromRaw};
 use system::Time;
 
 use csfml_audio_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 use std::ops::Deref;
 use inputstream::InputStream;
 use std::io::{Read, Seek};
