+++ src/graphics/font.rs
@@ -34,7 +34,7 @@ use csfml_graphics_sys as ffi;
 
 use std::io::{Read, Seek};
 use inputstream::InputStream;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Class for loading and manipulating character fonts
 pub struct Font {
