+++ src/graphics/image.rs
@@ -35,7 +35,7 @@ use csfml_graphics_sys as ffi;
 
 use std::io::{Read, Seek};
 use inputstream::InputStream;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Loading, manipulating and saving images.
 pub struct Image {
