+++ src/graphics/rectangle_shape.rs
@@ -32,7 +32,7 @@ use graphics::{Drawable, Shape, Transformable, FloatRect, IntRect, Color, Textur
 
 use csfml_system_sys::{sfBool, sfTrue, sfVector2f};
 use csfml_graphics_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Specialized shape representing a rectangle
 pub struct RectangleShape<'s> {
