+++ src/graphics/shader.rs
@@ -40,7 +40,7 @@ use csfml_graphics_sys as ffi;
 
 use std::io::{Read, Seek};
 use inputstream::InputStream;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Shader class (vertex and fragment)
 ///
