+++ src/lib.rs
@@ -96,12 +96,10 @@ extern crate csfml_system_sys;
 extern crate csfml_window_sys;
 extern crate csfml_graphics_sys;
 extern crate csfml_audio_sys;
-extern crate csfml_network_sys;
 
 mod inputstream;
 mod ext {
     pub mod event;
-    pub mod sf_bool_ext;
 }
 mod unicode_conv;
 
@@ -109,4 +107,3 @@ pub mod system;
 pub mod window;
 pub mod audio;
 pub mod graphics;
-pub mod network;
