+++ src/system/mod.rs
@@ -32,6 +32,7 @@ pub use self::sleep::sleep;
 pub use self::time::{Time, ZERO as TIME_ZERO};
 pub use self::clock::Clock;
 pub use self::sf_bool::{SfBool, TRUE as SF_TRUE, FALSE as SF_FALSE};
+use csfml_system_sys::{self, sfBool, sfTrue, sfFalse};
 
 mod time;
 mod clock;
@@ -40,3 +41,21 @@ mod vector2;
 mod vector3;
 mod sf_bool;
 pub mod raw_conv;
+
+#[doc(hidden)]
+pub trait SfBoolExt {
+    fn to_bool(self) -> bool;
+    fn from_bool(src: bool) -> Self;
+}
+
+impl SfBoolExt for sfBool {
+    fn to_bool(self) -> bool {
+        match self {
+            csfml_system_sys::sfFalse => false,
+            _ => true,
+        }
+    }
+    fn from_bool(src: bool) -> Self {
+        if src { sfTrue } else { sfFalse }
+    }
+}
