+++ src/window/keyboard.rs
@@ -22,7 +22,7 @@
 //
 
 use csfml_window_sys as ffi;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 
 /// Key codes known to SFML.
 #[repr(i32)]
