+++ src/window/sensor.rs
@@ -36,7 +36,7 @@
 
 use system::Vector3f;
 use csfml_window_sys::*;
-use ext::sf_bool_ext::SfBoolExt;
+use system::SfBoolExt;
 use system::raw_conv::{Raw, FromRaw};
 
 /// Get the current sensor value.
